requirements.txt
boto3==1.34.1
requests==2.31.0
pydantic==2.5.3
python-gitlab==4.3.0
jinja2==3.1.3
aws-lambda-powertools==2.30.2
python-jose[cryptography]==3.3.0 